LPS-146644 Move the PUBLICATIONS_USER role initialzation to InviteUsersMVCResourceCommand where it is going to be assigned to users
